Move the encryption code to a separate file
